refactor: rename SceneObjectRouter to SceneEventRouter, absorb 4 handlers
Extend the existing SceneObjectRouter (renamed to SceneEventRouter) with
GameObjectSpawned/Despawned and ItemSpawned/ItemDespawned handlers that
were pure pass-throughs in the plugin class. Authoring-log side-effects
are delegated via a Consumer<String> callback wired in startUp().

Removes 60 lines from CollectionLogHelperPlugin (2146 -> 2086).
